Moved GOLoadThread to the loader subdirectory
