+++ src/grandorgue/GOOrganController.cpp
@@ -42,6 +42,7 @@
 #include "gui/GOGUIPanelCreator.h"
 #include "gui/GOGUIRecorderPanel.h"
 #include "gui/GOGUISequencerPanel.h"
+#include "loader/GOLoadThread.h"
 #include "midi/GOMidi.h"
 #include "midi/GOMidiEvent.h"
 #include "midi/GOMidiPlayer.h"
@@ -67,7 +68,6 @@
 #include "GOFile.h"
 #include "GOFilename.h"
 #include "GOHash.h"
-#include "GOLoadThread.h"
 #include "GOMetronome.h"
 #include "GOOrgan.h"
 #include "GOPath.h"
